requirements.txt
numpy~=1.24.2
pandas~=1.5.3
PuLP~=2.7.0
google-cloud-storage~=2.7.0
